matricula/requirement.txt
django
django-simple-email-confirmation
django-ckeditor
djangoajax